+++ src/data/certificates.ts
@@ -184,7 +184,7 @@ const certificatesWithMetadata = Object.entries(certificateImages)
     
     return {
       id: filename,
-      image: `/assets/certificates/${path.split('/').pop()}`,
+      image: `assets/certificates/${path.split('/').pop()}`,
       ...metadata,
     } as Certificate;
   })